Moved UK to census folder, split metadata
