+++ tasks/uk/census/ons.py
@@ -1,38 +1,16 @@
 # http://webarchive.nationalarchives.gov.uk/20160105160709/http://ons.gov.uk/ons/guide-method/census/2011/census-data/bulk-data/bulk-data-downloads/index.html
 
-from luigi import Task, Parameter, LocalTarget
-
-from tasks.meta import OBSColumn, DENOMINATOR, current_session, OBSTag
-from tasks.util import (TableTask, ColumnsTask, classpath, shell,
-                        DownloadUnzipTask, TagsTask, TempTableTask, MetaWrapper, create_temp_schema)
-from tasks.uk.cdrc import OutputAreaColumns, OutputAreas
-from tasks.tags import UnitTags, SectionTags, SubsectionTags, LicenseTags
-
 from collections import OrderedDict
 import os
 import shutil
-import json
 from zipfile import ZipFile
 
+from luigi import Task, Parameter, LocalTarget
 
-def load_definition():
-    with open(os.path.join(os.path.dirname(__file__), 'columns.json')) as json_file:
-        return json.load(json_file, object_pairs_hook=OrderedDict)
-
-
-COLUMNS_DEFINITION = load_definition()
-
-
-class SourceTags(TagsTask):
-
-    def version(self):
-        return 1
-
-    def tags(self):
-        return [OBSTag(id='ons',
-                       name='Office for National Statistics (ONS)',
-                       type='source',
-                       description="The UK's largest independent producer of official statistics and the recognised national statistical institute of the UK (`ONS <https://www.ons.gov.uk/>`_)")]
+from tasks.meta import current_session
+from tasks.util import TableTask, classpath, shell, DownloadUnzipTask, TempTableTask, MetaWrapper, create_temp_schema
+from tasks.uk.cdrc import OutputAreaColumns, OutputAreas
+from tasks.uk.census.metadata import CensusColumns
 
 
 class DownloadEnglandWalesLocal(DownloadUnzipTask):
@@ -78,44 +56,6 @@ class ImportEnglandWalesLocal(TempTableTask):
         session.execute('ALTER TABLE {output} ADD PRIMARY KEY (geographycode)'.format(
             output=self.output().table
         ))
-
-
-class CensusColumns(ColumnsTask):
-
-    def requires(self):
-        return {
-            'units': UnitTags(),
-            'sections': SectionTags(),
-            'subsections': SubsectionTags(),
-            'source': SourceTags(),
-            'license': LicenseTags(),
-        }
-
-    def version(self):
-        return 4
-
-    def columns(self):
-        input_ = self.input()
-        source = input_['source']['ons']
-        license = input_['license']['uk_ogl']
-        uk = input_['sections']['uk']
-        subsections = input_['subsections']
-        units = input_['units']
-
-        columns = OrderedDict()
-        for key, column in COLUMNS_DEFINITION.items():
-            columns[key] = OBSColumn(
-                id=column['id'],
-                name=column['name'],
-                description='',
-                type='Numeric',
-                weight=column['weight'],
-                aggregate='sum',
-                targets={columns[denom]: DENOMINATOR for denom in column['denominators']},
-                tags=[uk, source, license, units[column['units']], subsections[column['subsection']]]
-            )
-
-        return columns
 
 
 class ImportAllEnglandWalesLocal(Task):
